moved CargoState enum definition to Constants.java
