+++ src/main/java/frc/robot/RobotContainer.java
@@ -15,6 +15,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.RunCommand;
+import frc.robot.Constants.CargoState;
 import frc.robot.RobotMap.mapControllers;
 import frc.robot.RobotPreferences.prefDrivetrain;
 import frc.robot.RobotPreferences.prefPreset;
@@ -160,10 +161,6 @@ public class RobotContainer {
     autoChooser.setDefaultOption("null", null);
     autoChooser.addOption("Four Ball A", autoFourBallA);
     SmartDashboard.putData(autoChooser);
-  }
-
-  public enum CargoState {
-    SHOOTING, COLLECTING, DISCARDING, PROCESSING, NONE
   }
 
   public Command getAutonomousCommand() {
