+++ src/main/java/frc/robot/commands/Cargo/DiscardCargo.java
@@ -7,7 +7,7 @@ package frc.robot.commands.Cargo;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.RobotContainer;
 import frc.robot.RobotPreferences;
-import frc.robot.RobotContainer.CargoState;
+import frc.robot.Constants.CargoState;
 import frc.robot.RobotPreferences.prefTransfer;
 import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Transfer;
